First batch of updates to unit tests
- Remove content from __init__.py
- Change test_script_main to use pytest
- Update test_script_main to run on windows as well as linux

+++ tests/test_script_main.py
@@ -1,32 +1,41 @@
 # encoding: utf-8
 import os
-import unittest
+import sys
+import pytest
 from tempfile import NamedTemporaryFile
 from os import path as op
 
 from send2trash.__main__ import main as trash_main
-from tests.test_plat_other import HOMETRASH
+
+# Only import HOMETRASH on supported platforms
+if sys.platform != "win32":
+    from send2trash.plat_other import HOMETRASH
 
 
-class TestMainTrash(unittest.TestCase):
-    def setUp(self):
-        self.file = NamedTemporaryFile(dir=op.expanduser('~'), prefix='send2trash_test', delete=False)
-
-    def test_trash(self):
-        trash_main(['-v', self.file.name])
-        self.assertFalse(op.exists(self.file.name))
-
-    def test_no_args(self):
-        self.assertRaises(SystemExit, trash_main, [])
-        self.assertRaises(SystemExit, trash_main, ['-v'])
-        self.assertTrue(op.exists(self.file.name))
-        trash_main([self.file.name])  # Trash the file so tearDown runs properly
-
-    def tearDown(self):
-        name = op.basename(self.file.name)
-        os.remove(op.join(HOMETRASH, 'files', name))
-        os.remove(op.join(HOMETRASH, 'info', name + '.trashinfo'))
+@pytest.fixture
+def file():
+    file = NamedTemporaryFile(
+        dir=op.expanduser("~"), prefix="send2trash_test", delete=False
+    )
+    file.close()
+    yield file.name
+    # Cleanup trash files on supported platforms
+    if sys.platform != "win32":
+        name = op.basename(file.name)
+        # Remove trash files if they exist
+        if op.exists(op.join(HOMETRASH, "files", name)):
+            os.remove(op.join(HOMETRASH, "files", name))
+            os.remove(op.join(HOMETRASH, "info", name + ".trashinfo"))
+    if op.exists(file.name):
+        os.remove(file.name)
 
 
-if __name__ == '__main__':
-    unittest.main()
+def test_trash(file):
+    trash_main(["-v", file])
+    assert op.exists(file) is False
+
+
+def test_no_args(file):
+    pytest.raises(SystemExit, trash_main, [])
+    pytest.raises(SystemExit, trash_main, ["-v"])
+    assert op.exists(file) is True
